refactor: Expression 에서 사칙 연산 문자열 생성 기능 분리하기
